修改我的喜欢item界面,Room保存ShoppingCart

+++ core/designsystem/src/main/java/com/example/designsystem/common/AddAndRemoveFood.kt
@@ -8,19 +8,17 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.Remove
 import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.example.designsystem.theme.LocalTintTheme
-import com.example.model.remoteModel.Food
 
 @Composable
 fun AddAndRemoveFood(
     modifier: Modifier = Modifier,
-    num:Int,
+    num: Long,
     onRemove:() -> Unit,
     onAdd:() -> Unit
 ) {
